\name{chemist}
\docType{data}
\alias{chemist}

\title{Coordinates of chemist shops in Montpellier, France.}
\description{
  This data set gives the X and Y coordinates of the 99 chemist shops in Montpellier, France.
}
\usage{data(chemist)
}

\format{A data frame containing 99 coordinates}
\source{GPS location by Christophe Dematte{\"\i}
}
\keyword{datasets}
