\name{delai}
\alias{delai}

\title{Time computation between two dates}
\description{
  Computes the time between two dates.
}
\usage{
delai(date1, date2)
}

\arguments{
  \item{date1}{The first date, obtained by the function date().}
  \item{date2}{The second date, also obtained by the function date().}
}
\value{
  An output giving the date in hh:mm:ss format.
}
\author{Christophe Dematte{\"\i}  \email{demattei@iurc.montp.inserm.fr}
}
\keyword{spatial}
