\name{dist2p}
\alias{dist2p}

\title{Euclidian distance computation in 2D}
\description{
  Computes the euclidian distance between two 2D point coordinates or between one point and a vector of points.
}
\usage{
dist2p(x, y, x0, y0)
}

\arguments{
  \item{x}{X-coordinate of the second point or the vector of point}
  \item{y}{Y-coordinate of the second point or the vector of point}
  \item{x0}{X-coordinate of the first point}
  \item{y0}{Y-coordinate of the first point}
}
\value{
  \item{dist}{The distance computed}
}
\author{Christophe Dematte{\"\i}  \email{demattei@iurc.montp.inserm.fr}
}
\keyword{spatial}

