\name{dist2p3d}
\alias{dist2p3d}

\title{Euclidian distance computation in 3D}
\description{
  Computes the euclidian distance between two 3D point coordinates or between one point and a vector of points.
}
\usage{
dist2p3d(x, y, z, x0, y0, z0)
}

\arguments{
  \item{x}{X-coordinate of the second point or the vector of point}
  \item{y}{Y-coordinate of the second point or the vector of point}
  \item{z}{Z-coordinate of the second point or the vector of point}
  \item{x0}{X-coordinate of the first point}
  \item{y0}{Y-coordinate of the first point}
  \item{z0}{Z-coordinate of the first point}
}
\value{
  \item{dist}{The distance computed}
}
\author{Christophe Dematte{\"\i}  \email{demattei@iurc.montp.inserm.fr}
}
\keyword{spatial}

