\name{grille}
\docType{data}
\alias{grille}

\title{Underlying population coordinates}
\description{
  This data set gives the X and Y coordinates of 2251 underlying population individuals. It takes the Montpellier density into account. Each point represents $100$ inhabitants.
}
\usage{data(grille)
}

\format{A $2$ column matrix containing 2251 coordinates}
\source{French 1999 population census, INSEE data.
}
\references{
Pallez, D. 2000. Recensement de la population 1999 - Quartiers de Montpellier.  \emph{Rep\`eres pour l'Economie du Languedoc-Roussillon, Synth\`ese, INSEE Languedoc-Roussilon}, \bold{4}, 1--6.
}
\keyword{datasets}
