\name{irislist}
\docType{data}
\alias{irislist}

\title{List of the $30$ Montpellier IRIS bounds.}
\description{
  Montpellier is divided in $30$ quarters, called IRIS. This data sets is a list in wich each objects is an object of class "owin" giving the IRIS bounds. The "owin" class is defined in the "spatstat" R package.
}
\usage{data(irislist)
}

\format{A list of $30$ objects of class "owin".}
\source{GPS location by Christophe Dematte\"\i
}
\references{CDrom INSEE - R\'ef TABILOIRI. Correspondances IRIS-\^ilots, Recensement de la population 1999. 14/11/2000.}
\keyword{datasets}
